Upload: deleting file on the server

+++ 08_working-with-express/12_authentication/controllers/admin.js
@@ -1,4 +1,5 @@
 const { validationResult } = require('express-validator/check');
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -144,6 +145,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save()
@@ -181,11 +183,17 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
-    })
+
+  Product.findById(prodId).then(product => {
+    if (!product) {
+      return next(new Error('Product not found'));
+    }
+    fileHelper.deleteFile(product.imageUrl);
+    return Product.deleteOne({ _id: prodId, userId: req.user._id });
+  }).then(() => {
+    console.log('DESTROYED PRODUCT');
+    res.redirect('/admin/products');
+  })
     .catch(err => {
       const error = new Error(err);
       error.httpStatusCode = 500;
